fixing bug 8159: it-drop and thiefs
Note that the fixing is special to this purpose, there is still
some severe problem in the code:
When an actor activates a thief, it remembers the actor by
memory-address. Then it plays its animation and later check the
actor's properties. When the actor got killed in the mean-time
(like it happens to the it-rotor that is created by it-drop),
this check will cause a segfault.

This fix checks the entering actor and only the four marble-
kind actors further activate the thief. it-rotor does not.
This way, it-drop becomes safe, but other circumstances in
which actors get killed, still suffer from a segfault here.